# swagger/sol_delilah.yaml
swagger: '2.0'
info:
  description: Delilah Restó is a food ordering app for restaurants
  version: 1.0.0
  title: Delilah Restó
  contact:
    email: solmartinezpisani@gmail.com
...
schemes:
  - https
  - http
paths:
...
  /orders:
    get:
      tags:
        - orders
      summary: Get all orders (admin)
      operationId: getOrders
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
      responses:
        '200':
          description: Successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/GetOrders'
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404
  /orders/pending:
    get:
      tags:
        - orders
      summary: Get all pending orders (admin)
      operationId: getPending
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
      responses:
        '200':
          description: Successful operation
          schema:
            type: array
            items:
              type: object
              properties:
                order_date:
                  type: string
                  example: '2021-01-27T14:34:01.000Z'
                id:
                  type: integer
                  format: int64
                  example: 13
                total_price:
                  type: integer
                  format: int64
                  example: 300
                user_id:
                  type: "integer"
                  format: "int64"
                  example: 1
                name:
                  type: string
                  example: "Sol Martinez"
                address:
                  type: string
                  example: "Cornelia Street 13"
                status:
                  type: string
                  example: "Pending"
                description:
                  type: array
                  items:
                    type: string
                    example: "2x Pizza"
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404
  /orders/inprogress:
    get:
      tags:
        - orders
      summary: Get all the orders that are in progress (admin)
      operationId: getInProgress
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
      responses:
        '200':
          description: Successful operation
          schema:
            type: array
            items:
              type: object
              properties:
                order_date:
                  type: string
                  example: '2021-01-27T14:34:01.000Z'
                id:
                  type: integer
                  format: int64
                  example: 13
                total_price:
                  type: integer
                  format: int64
                  example: 300
                user_id:
                  type: integer
                  format: int64
                  example: 1
                name:
                  type: string
                  example: "Sol Martinez"
                address:
                  type: string
                  example: "Cornelia Street 13"
                status:
                  type: string
                  example: "In Progress"
                description:
                  type: array
                  items:
                    type: string
                    example: "2x Pizza"
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404
...
  /orders/shipped:
    get:
      tags:
        - orders
      summary: Get all shipped orders (admin)
      operationId: getShippedOrders
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
      responses:
        '200':
          description: Successful operation
          schema:
            type: array
            items:
              type: object
              properties:
                order_date:
                  type: string
                  example: '2021-01-27T14:34:01.000Z'
                id:
                  type: integer
                  format: int64
                  example: 13
                total_price:
                  type: integer
                  format: int64
                  example: 300
                user_id:
                  type: string
                  example: '1'
                name:
                  type: string
                  example: Sol Martinez
                address:
                  type: string
                  example: Cornelia Street 13
                status:
                  type: string
                  example: Shipped
                description:
                  type: array
                  items:
                    type: string
                    example: "2x Pizza"
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404
...
  /orders/cancelled:
    get:
      tags:
        - orders
      summary: Get all cancelled orders (admin)
      operationId: getCancelledOrders
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
      responses:
        '200':
          description: Successful operation
          schema:
            type: array
            items:
              type: object
              properties:
                order_date:
                  type: string
                  example: '2021-01-27T14:34:01.000Z'
                id:
                  type: integer
                  format: int64
                  example: 13
                total_price:
                  type: integer
                  format: int64
                  example: 300
                user_id:
                  type: "integer"
                  format: "int64"
                  example: 1
                name:
                  type: string
                  example: "Sol Martinez"
                address:
                  type: string
                  example: "Cornelia Street 13"
                status:
                  type: string
                  example: "Cancelled"
                description:
                  type: array
                  items:
                    type: string
                    example: "2x Pizza"
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404
  /orders/{id}:
    patch:
      tags:
        - orders
      summary: Updates an order status
      operationId: updateOrder
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the order to update
        - in: body
          name: body
          description: |-
            Update order string. It has 8 possible values: 
             1 (equals to Cart),  
             2 (equals to Checkout),  
             3 (equals to Pending),  
             4 (equals to In Progress),  
             5 (equals to Ready),  
              6 (equals to Shipped),  
             7 (equals to Delivered),  
             8 (equals to Cancelled)
          schema:
            type: object
            properties:
              idStatus:
                type: integer
                format: int64
                example: '3'
      responses:
        '200':
          description: Successful operation
          schema:
            $ref: '#/definitions/UpdateOrderResponse'
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404
    delete:
      tags:
        - orders
      summary: Delete order
      operationId: deleteOrder
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the order to delete.
      responses:
        '200':
          description: Successful operation
          schema:
            $ref: '#/definitions/DeleteOrderResponse'
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404
definitions:
  Authenticate:
    type: object
    properties:
      message:
        type: string
        example: Header must contain JWT to verify identity.
...
  GetOrders:
    type: object
    properties:
      order_date:
        type: string
        example: '2021-01-27T14:34:01.000Z'
      id:
        type: integer
        format: int64
        example: 13
      total_price:
        type: integer
        format: int64
        example: 300
      user_id:
        type: string
        example: '1'
      name:
        type: string
        example: Sol Martinez
      address:
        type: string
        example: Cornelia Street 13
      status:
        type: string
        example: Cart
      description:
        type: array
        items:
          type: string
          example: 2x Pizza
...
  DeleteOrderResponse:
    type: object
    properties:
      message:
        type: string
        example: Order deleted successfully
  UpdateOrderResponse:
    type: object
    properties:
      message:
        type: string
        example: Order status changed successfully
...